add API integration and data loading for "All Users" tab on main screen

+++ TMessagesProj/src/main/java/org/telegram/divo/screen/profile/ProfileViewState.kt
@@ -66,10 +66,10 @@ data class SocialLinks(
 data class PhysicalParams(
     val gender: String = "",
     val age: String = "",
-    val height: Int = 0,
+    val height: Float = 0f,
     val waist: Int = 0,
     val hips: Int = 0,
-    val shoeSize: Int = 0,
+    val shoeSize: Float = 0f,
     val hairLength: String = "",
     val hairColor: String = "",
     val eyeColor: String = "",
